fix model

+++ database/migrations/2020_09_28_235534_create_table_course_user.php
@@ -14,9 +14,9 @@ class CreateTableCourseUser extends Migration
     public function up()
     {
         Schema::create('user_course', function (Blueprint $table) {
-            $table->bigIncrements('id_user_course');
-            $table->unsignedBigInteger('id_user');
-            $table->unsignedBigInteger('id_course');
+            $table->increments('id_user_course');
+            $table->unsignedInteger('id_user');
+            $table->integer('id_course');
             $table->foreign('id_course')->references('id_course')->on('course');
             $table->string("rol");
             $table->string("state");
